Deleted LICENSE
I’m not sure if we’ve agreed on a license to use for this project at the moment so I’ve deleted the GPL license with which I initialized the project. For the time being I’ve noted in the README that this software is owned by our three contributors and that all rights are reserved. We can discuss adopting an open-source license in the near future.